+++ packages/twenty-server/src/modules/calendar/jobs/blocklist-reimport-calendar-events.job.ts
@@ -3,7 +3,7 @@ import { Logger, Scope } from '@nestjs/common';
 import { Processor } from 'src/engine/integrations/message-queue/decorators/processor.decorator';
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
-import { GoogleCalendarSyncService } from 'src/modules/calendar/services/calendar-sync-manager/google-calendar-sync.service';
+import { GoogleCalendarSyncService } from 'src/modules/calendar/services/calendar-event-import-manager/google-calendar-sync.service';
 import { ConnectedAccountRepository } from 'src/modules/connected-account/repositories/connected-account.repository';
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 import { Process } from 'src/engine/integrations/message-queue/decorators/process.decorator';
