+++ packages/twenty-server/src/modules/connected-account/auto-companies-and-contacts-creation/auto-companies-and-contacts-creation.module.ts
@@ -8,9 +8,9 @@ import { ObjectMetadataRepositoryModule } from 'src/engine/object-metadata-repos
 import { PersonWorkspaceEntity } from 'src/modules/person/standard-objects/person.workspace-entity';
 import { WorkspaceMemberWorkspaceEntity } from 'src/modules/workspace-member/standard-objects/workspace-member.workspace-entity';
 import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
-import { CalendarEventParticipantModule } from 'src/modules/calendar/services/calendar-event-participant/calendar-event-participant.module';
 import { FeatureFlagEntity } from 'src/engine/core-modules/feature-flag/feature-flag.entity';
 import { MessagingCommonModule } from 'src/modules/messaging/common/messaging-common.module';
+import { CalendarEventParticipantModule } from 'src/modules/calendar/services/calendar-event-participant-manager/calendar-event-participant.module';
 
 @Module({
   imports: [
